refactor(edit-engine): extract Cell reset execution
Move clear drawing, reset placement, and reset body mutation bodies behind a typed Cell reset dispatcher. The transaction coordinator continues to own ordered execution, deferred Net pruning, revision/status updates, and final validation while the dispatcher reports only connectivity and geometry outcomes.

Add a permanent public transaction boundary regression proving 1024 edits execute atomically and the 1025th is rejected before execution. Existing Cell reset planner/history tests continue to cover retained interface and undo semantics.

Validation: pnpm test:local packages/edit-engine/src (246 tests); pnpm typecheck; git diff --check.

Gate selection: shared edit-engine hotspot; focused package tests during extraction, hierarchy and full-delivery gates deferred to the completed migration branch.

Unrelated user-owned probe-conflicts.mjs in the main checkout remains untouched.

Test-Impact: tests-updated
