+++ packages/edit-engine/src/transaction.ts
@@ -50,6 +50,7 @@ import {
 } from "./transaction-instance-annotations.js";
 import { reconcileTransformDirectContacts } from "./transaction-direct-contact.js";
 import { nextPhysicalContactOperation } from "./transaction-connectivity-normalizer.js";
+import { applyCellResetEdit } from "./transaction-cell-reset.js";
 import { applyCellInterfaceEdit } from "./transaction-cell-interface.js";
 import { applyPresentationLayoutEdit } from "./transaction-presentation-layout.js";
 import {
@@ -259,144 +260,16 @@ export function executeTransaction(
       case "undo":
       case "redo":
         continue;
-      case "clear_cell_drawing": {
-        const routeIds = new Set(draft.routes.map((route) => route.id));
-        const ownerNetIds = removeConnectivityEvidenceOwnedBy(
-          draft,
-          routeIds,
-          changedObjectIds,
-        );
-        for (const object of [
-          ...draft.routes,
-          ...(draft.drafting?.objects ?? []),
-        ]) {
-          changedObjectIds.add(object.id);
-        }
-        draft.routes = [];
-        draft.drafting = { objects: [] };
-        for (const netId of ownerNetIds) {
-          deferNetPrune(netId);
-        }
-        if (ownerNetIds.length > 0) connectivityChanged = true;
-        geometryChanged = true;
-        break;
-      }
-      case "reset_cell_placement": {
-        const routeIds = new Set(draft.routes.map((route) => route.id));
-        const ownerNetIds = removeConnectivityEvidenceOwnedBy(
-          draft,
-          routeIds,
-          changedObjectIds,
-        );
-        for (const object of [
-          ...draft.instances.filter((instance) => instance.placement !== null),
-          ...draft.routes,
-          ...draft.layoutGroups,
-          ...draft.constraints,
-        ]) {
-          changedObjectIds.add(object.id);
-        }
-        for (const instance of draft.instances) instance.placement = null;
-        draft.routes = [];
-        draft.layoutGroups = [];
-        draft.constraints = [];
-        for (const netId of ownerNetIds) {
-          deferNetPrune(netId);
-        }
-        if (ownerNetIds.length > 0) connectivityChanged = true;
-        geometryChanged = true;
-        break;
-      }
+      case "clear_cell_drawing":
+      case "reset_cell_placement":
       case "reset_cell_body": {
-        const cellPinInstanceIds = new Set(
-          draft.netlist?.terminals.flatMap(
-            (terminal) => terminal.interfaceInstanceIds,
-          ) ?? [],
-        );
-        const interfaceNetIds = new Set(
-          draft.netlist?.terminals.map((terminal) => terminal.netId) ?? [],
-        );
-        const retainedInstances = draft.instances.filter((instance) =>
-          cellPinInstanceIds.has(instance.id),
-        );
-        const retainedNets = draft.nets
-          .filter((net) => interfaceNetIds.has(net.id))
-          .map((net) => ({
-            ...net,
-            terminals: net.terminals.filter((terminal) =>
-              cellPinInstanceIds.has(terminal.instanceId),
-            ),
-          }));
-        const retainedAnnotations = draft.annotations.filter(
-          (annotation) =>
-            annotation.anchor.kind === "object" &&
-            cellPinInstanceIds.has(annotation.anchor.objectId),
-        );
-        const retainedAnnotationIds = new Set(
-          retainedAnnotations.map((annotation) => annotation.id),
-        );
-        const retainedEvidence = draft.connectivityEvidence.filter(
-          (evidence) => {
-            if (evidence.kind === "explicit-equivalence") {
-              return evidence.memberNetIds.every((netId) =>
-                interfaceNetIds.has(netId),
-              );
-            }
-            if (!interfaceNetIds.has(evidence.netId)) return false;
-            if (evidence.kind !== "name-claim") return true;
-            switch (evidence.owner.kind) {
-              case "explicit-net-property":
-                return true;
-              case "net-label":
-                return retainedAnnotationIds.has(evidence.owner.annotationId);
-              case "power-marker":
-                return (
-                  cellPinInstanceIds.has(evidence.owner.objectId) ||
-                  retainedAnnotationIds.has(evidence.owner.objectId)
-                );
-            }
-          },
-        );
-        const retainedIds = new Set([
-          ...retainedInstances.map((instance) => instance.id),
-          ...retainedNets.map((net) => net.id),
-          ...retainedAnnotations.map((annotation) => annotation.id),
-          ...retainedEvidence.map((evidence) => evidence.id),
-        ]);
-        for (const object of [
-          ...draft.instances,
-          ...draft.nets,
-          ...draft.routes,
-          ...draft.junctions,
-          ...draft.noConnects,
-          ...draft.annotations,
-          ...draft.connectivityEvidence,
-          ...draft.layoutGroups,
-          ...draft.constraints,
-          ...(draft.drafting?.objects ?? []),
-        ]) {
-          if (!retainedIds.has(object.id)) changedObjectIds.add(object.id);
-        }
-        for (const retainedNet of retainedNets) {
-          const sourceNet = draft.nets.find((net) => net.id === retainedNet.id);
-          if (sourceNet?.terminals.length !== retainedNet.terminals.length) {
-            changedObjectIds.add(retainedNet.id);
-          }
-        }
-        draft.instances = retainedInstances;
-        draft.nets = retainedNets;
-        draft.routes = [];
-        draft.junctions = [];
-        draft.noConnects = [];
-        draft.annotations = retainedAnnotations;
-        draft.connectivityEvidence = retainedEvidence;
-        draft.layoutGroups = [];
-        draft.constraints = [];
-        draft.drafting = { objects: [] };
-        if (draft.mosBulkDefaults) changedObjectIds.add(draft.id);
-        delete draft.mosBulkDefaults;
-        connectivityChanged = true;
-        geometryChanged = true;
+        const outcome = applyCellResetEdit(edit, {
+          draft,
+          changedObjectIds,
+          deferNetPrune,
+        });
+        connectivityChanged ||= outcome.connectivityChanged;
+        geometryChanged ||= outcome.geometryChanged;
         break;
       }
       case "add_instance": {
